fix: fix scroll to top of long emails in view by conversation
