+++ src/views/app/detail-panel.tsx
@@ -9,7 +9,7 @@ import { Container } from '@zextras/carbonio-design-system';
 import { useAppContext } from '@zextras/carbonio-shell-ui';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
 
-import { ConversationPreviewPanel } from './detail-panel/conversation-preview-panel';
+import { ConversationPreviewPanelContainer } from './detail-panel/conversation-preview-panel-container';
 import { MessagePreviewPanelContainer } from './detail-panel/message-preview-panel-container';
 import { SelectionInteractive } from './detail-panel/selection-interactive';
 import type { AppContext } from '../../types';
@@ -24,7 +24,7 @@ const DetailPanel: FC = () => {
 					<SelectionInteractive count={count} />
 				</Route>
 				<Route exact path={`${path}/folder/:folderId/conversation/:conversationId`}>
-					<ConversationPreviewPanel />
+					<ConversationPreviewPanelContainer />
 				</Route>
 				<Route exact path={`${path}/folder/:folderId/message/:messageId`}>
 					<MessagePreviewPanelContainer />
