+++ src/views/app/detail-panel/conversation-message-preview.tsx
@@ -10,22 +10,22 @@ import { Padding } from '@zextras/carbonio-design-system';
 import MailPreview from './preview/mail-preview';
 import { useMessageActions } from '../../../hooks/use-message-actions';
 import { MailMessage } from '../../../types';
-import { useExtraWindow } from '../extra-windows/use-extra-window';
 
 export type ConversationMessagePreviewProps = {
 	message: MailMessage;
 	idPrefix: string;
 	isExpanded: boolean;
 	isAlone: boolean;
+	isInsideExtraWindow: boolean;
 };
 
 export const ConversationMessagePreview: FC<ConversationMessagePreviewProps> = ({
 	idPrefix,
 	message,
 	isExpanded,
-	isAlone
+	isAlone,
+	isInsideExtraWindow
 }) => {
-	const { isInsideExtraWindow } = useExtraWindow();
 	const messageActions = useMessageActions(message, isAlone);
 	return (
 		<Padding key={`${idPrefix}-${message.id}`} bottom="medium" width="100%">
