+++ src/views/app/detail-panel/conversation-preview-panel-container.tsx
@@ -3,84 +3,23 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import React, { FC, ReactElement, useCallback, useEffect, useMemo } from 'react';
+import React, { FC, useEffect, useMemo } from 'react';
 
-import { Container, Shimmer } from '@zextras/carbonio-design-system';
-import { FOLDERS, useTags, useUserSettings } from '@zextras/carbonio-shell-ui';
-import { filter, find, map, sortBy } from 'lodash';
+import { Container } from '@zextras/carbonio-design-system';
+import { FOLDERS, useTags } from '@zextras/carbonio-shell-ui';
+import { filter, find } from 'lodash';
 import { useParams } from 'react-router-dom';
 
-import { ConversationMessagePreview } from './conversation-message-preview';
+import { ConversationPreviewPanel } from './conversation-preview-component';
 import PreviewPanelHeader from './preview/preview-panel-header';
 import { useAppDispatch, useAppSelector } from '../../../hooks/redux';
 import { getConv, searchConv } from '../../../store/actions';
 import {
 	selectConversationExpandedStatus,
-	selectConversationsArray,
-	selectCurrentFolderExpandedStatus
+	selectConversationsArray
 } from '../../../store/conversations-slice';
-import { selectMessages } from '../../../store/messages-slice';
 import type { Conversation, MailsStateType } from '../../../types';
 import { useExtraWindow } from '../extra-windows/use-extra-window';
-
-export const ConversationPreviewComponent = ({
-	conversation
-}: {
-	conversation: Conversation;
-}): ReactElement => {
-	const settings = useUserSettings();
-	const messages = useAppSelector(selectMessages);
-	const conversationStatus = useAppSelector(selectCurrentFolderExpandedStatus)[conversation.id];
-	const convMessages = useMemo(() => {
-		const msgs = map(conversation?.messages, (item) => messages[item.id] ?? item);
-
-		if (settings.prefs.zimbraPrefConversationOrder === 'dateAsc' && msgs?.length > 0) {
-			return sortBy(msgs, [(o): string | number => o.date]);
-		}
-		return msgs ?? [];
-	}, [conversation?.messages, settings.prefs.zimbraPrefConversationOrder, messages]);
-
-	const isExpanded = useCallback(
-		(index: number): boolean => {
-			if (settings.prefs.zimbraPrefConversationOrder === 'dateAsc') {
-				return index === convMessages.length - 1;
-			}
-			return index === 0;
-		},
-		[convMessages.length, settings.prefs.zimbraPrefConversationOrder]
-	);
-
-	return (
-		<Container
-			style={{ overflowY: 'auto' }}
-			height="fill"
-			background="gray5"
-			padding={{ horizontal: 'large', bottom: 'small', top: 'large' }}
-			mainAlignment="flex-start"
-		>
-			<Container height="fit" mainAlignment="flex-start" background="gray5">
-				{conversation && conversationStatus === 'complete' ? (
-					<>
-						{map(convMessages, (message, index) =>
-							message ? (
-								<ConversationMessagePreview
-									idPrefix={conversation.id}
-									message={message}
-									isExpanded={isExpanded(index)}
-									isAlone={convMessages?.length === 1}
-								></ConversationMessagePreview>
-							) : (
-								<Shimmer.Logo size="large" />
-							)
-						)}
-					</>
-				) : (
-					<></>
-				)}
-			</Container>
-		</Container>
-	);
-};
 
 type ConversationPreviewPanelProps = { conversationId?: string; folderId?: string };
 
@@ -94,7 +33,7 @@ const useConversationPreviewPanelParameters = (
 	};
 };
 
-export const ConversationPreviewPanel: FC<ConversationPreviewPanelProps> = (props) => {
+export const ConversationPreviewPanelContainer: FC<ConversationPreviewPanelProps> = (props) => {
 	const { conversationId, folderId } = useConversationPreviewPanelParameters(props);
 	const tagsFromStore = useTags();
 	const { isInsideExtraWindow } = useExtraWindow();
@@ -137,7 +76,10 @@ export const ConversationPreviewPanel: FC<ConversationPreviewPanelProps> = (prop
 			{showPreviewPanel && (
 				<>
 					{!isInsideExtraWindow && <PreviewPanelHeader item={conversation} folderId={folderId} />}
-					<ConversationPreviewComponent conversation={conversation} />
+					<ConversationPreviewPanel
+						conversation={conversation}
+						isInsideExtraWindow={isInsideExtraWindow}
+					/>
 				</>
 			)}
 		</Container>
